Make test suite PHP 8 compatible

+++ tests/BaseTests.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace Dwoo\Tests
-{
+namespace Dwoo\Tests {
 
     use Dwoo\Compiler;
     use Dwoo\Core;
-    use PHPUnit_Framework_Constraint;
-    use PHPUnit_Framework_TestCase;
+    use PHPUnit\Framework\Constraint\Constraint;
+    use PHPUnit\Framework\TestCase;
 
     /**
      * Class BaseTests
      *
      * @package Dwoo\Tests
      */
-    abstract class BaseTests extends PHPUnit_Framework_TestCase
+    abstract class BaseTests extends TestCase
     {
 
         protected $compileDir;
@@ -25,17 +24,17 @@ namespace Dwoo\Tests
          * Constructs a test case with the given name.
          *
          * @param string $name
-         * @param array  $data
+         * @param array $data
          * @param string $dataName
          */
         public function __construct($name = null, array $data = array(), $dataName = '')
         {
             $this->compileDir = __DIR__ . DIRECTORY_SEPARATOR . 'temp' . DIRECTORY_SEPARATOR . 'cache';
-            $this->cacheDir   = __DIR__ . DIRECTORY_SEPARATOR . 'temp' . DIRECTORY_SEPARATOR . 'compiled';
+            $this->cacheDir = __DIR__ . DIRECTORY_SEPARATOR . 'temp' . DIRECTORY_SEPARATOR . 'compiled';
 
             // extend this class and override this in your constructor to test a modded compiler
             $this->compiler = new Compiler();
-            $this->dwoo     = new Core($this->compileDir, $this->cacheDir);
+            $this->dwoo = new Core($this->compileDir, $this->cacheDir);
 
             parent::__construct($name, $data, $dataName);
         }
@@ -51,25 +50,27 @@ namespace Dwoo\Tests
 
         /**
          * @param string $path
-         * @param bool   $emptyOnly
+         * @param bool $emptyOnly
          */
         protected function clearDir($path, $emptyOnly = false)
         {
-            if (is_dir($path)) {
-                foreach (glob($path . '/*') as $f) {
-                    $this->clearDir($f);
+            if ($path !== null) {
+                if (is_dir($path)) {
+                    foreach (glob($path . '/*') as $f) {
+                        $this->clearDir($f);
+                    }
+                    if (!$emptyOnly) {
+                        rmdir($path);
+                    }
+                } elseif (is_file($path)) {
+                    unlink($path);
                 }
-                if (!$emptyOnly) {
-                    rmdir($path);
-                }
-            } elseif (is_file($path)) {
-                unlink($path);
             }
         }
     }
 
     // Evaluates two strings and ignores differences in line endings (\r\n == \n == \r)
-    class DwooConstraintStringEquals extends PHPUnit_Framework_Constraint
+    class DwooConstraintStringEquals extends Constraint
     {
         protected $target;
 
@@ -78,20 +79,20 @@ namespace Dwoo\Tests
             $this->target = preg_replace('#(\r\n|\r)#', "\n", $target);
         }
 
-        public function evaluate($other, $description = '', $returnResult = false)
+        public function evaluate($other, string $description = '', bool $returnResult = false): ?bool
         {
             $this->other = preg_replace('#(\r\n|\r)#', "\n", $other);
 
             return $this->target == $this->other;
         }
 
-        public function toString()
+        public function toString(): string
         {
             return 'equals expected value.' . PHP_EOL . 'Expected:' . PHP_EOL . $this->target . PHP_EOL . 'Received:' . PHP_EOL . $this->other . PHP_EOL;
         }
     }
 
-    class DwooConstraintPathEquals extends PHPUnit_Framework_Constraint
+    class DwooConstraintPathEquals extends Constraint
     {
         protected $target;
 
@@ -100,14 +101,14 @@ namespace Dwoo\Tests
             $this->target = preg_replace('#([\\\\/]{1,2})#', '/', $target);
         }
 
-        public function evaluate($other, $description = '', $returnResult = false)
+        public function evaluate($other, string $description = '', bool $returnResult = false): ?bool
         {
             $this->other = preg_replace('#([\\\\/]{1,2})#', '/', $other);
 
             return $this->target == $this->other;
         }
 
-        public function toString()
+        public function toString(): string
         {
             return 'equals expected value.' . PHP_EOL . 'Expected:' . PHP_EOL . $this->target . PHP_EOL . 'Received:' . PHP_EOL . $this->other . PHP_EOL;
         }
